Distribute test_variational_inference
