+++ pymc/tests/variational/test_callbacks.py
@@ -1,0 +1,54 @@
+#   Copyright 2020 The PyMC Developers
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+
+import aesara
+import numpy as np
+import pytest
+
+import pymc as pm
+
+
+@pytest.mark.parametrize("diff", ["relative", "absolute"])
+@pytest.mark.parametrize("ord", [1, 2, np.inf])
+def test_callbacks_convergence(diff, ord):
+    cb = pm.variational.callbacks.CheckParametersConvergence(every=1, diff=diff, ord=ord)
+
+    class _approx:
+        params = (aesara.shared(np.asarray([1, 2, 3])),)
+
+    approx = _approx()
+
+    with pytest.raises(StopIteration):
+        cb(approx, None, 1)
+        cb(approx, None, 10)
+
+
+def test_tracker_callback():
+    import time
+
+    tracker = pm.callbacks.Tracker(
+        ints=lambda *t: t[-1],
+        ints2=lambda ap, h, j: j,
+        time=time.time,
+    )
+    for i in range(10):
+        tracker(None, None, i)
+    assert "time" in tracker.hist
+    assert "ints" in tracker.hist
+    assert "ints2" in tracker.hist
+    assert len(tracker["ints"]) == len(tracker["ints2"]) == len(tracker["time"]) == 10
+    assert tracker["ints"] == tracker["ints2"] == list(range(10))
+    tracker = pm.callbacks.Tracker(bad=lambda t: t)  # bad signature
+    with pytest.raises(TypeError):
+        tracker(None, None, 1)
